+++ app/cmd/crypt.py
@@ -11,4 +11,3 @@ from flask_bcrypt import (
 def password_hash(password):
 
     password_hash = generate_password_hash(password.encode('utf-8')).decode('utf-8')
-    print(password_hash)