+++ frontend/src/config/menu.js
@@ -75,13 +75,28 @@ const menus = {
   resultats: {
     icon: "show_chart",
     label: "Résultats des suivis",
-    names: ["resultats.declarations", "resultats.in"]
+    names: [
+      "chasse.restitution_gd_public",
+      "resultats.declarations",
+      "resultats.in",
+      "chasse.restitution_indices_performances",
+      "icia.restitution_gd_public",
+      "degats_agricoles.restitution_gd_public",
+      "sylviculture.restitution_diagnostics_sylvicoles",
+      "sylviculture.restitution_suivi_peuplements_sensibles"
+    ]
   },
 
   chasse: {
     icon: "fa-cog",
     label: 'Chasse',
-    names: ['chasse.saisie', 'chasse.admin', 'chasse.bilan', 'chasse.exports', 'chasse.page_type'],
+    names: [
+      "chasse.saisie",
+      "chasse.admin",
+      "chasse.bilan",
+      "chasse.restitution_bilan_detaille",
+      "chasse.exports"
+    ],
     hidden: ({ $store }) => $store.getters.droitMax < 5
   },